fixing remove logic
